@@ -18,10 +18,10 @@ package android.example.com.squawker.following;
 import android.content.SharedPreferences;
 import android.example.com.squawker.R;
 import android.os.Bundle;
-import android.support.v7.preference.Preference;
-import android.support.v7.preference.PreferenceFragmentCompat;
-import android.support.v7.preference.SwitchPreferenceCompat;
 import android.util.Log;
+import androidx.preference.Preference;
+import androidx.preference.PreferenceFragmentCompat;
+import androidx.preference.SwitchPreferenceCompat;
 
 import com.google.firebase.messaging.FirebaseMessaging;
 
